Add support for reading fragment BED files with polars.
Add support for reading fragment BED files with polars.
For now this is only 35% faster, but this is mainly due
to the conversion to a PyRanges dataframe which always
converts "Chromosome" and "Strand" column to "str" and
then to categorical, even if it is already a "str" or
categorical column.
The "Name" column is converted to categorical as later
in the code several groupby operations are done on that
column.