Add "create_regions_topics_frequency_matrix" function to replace "load_word_topics" in the future.
Add "create_regions_topics_frequency_matrix" function, which uses Mallet region topics
counts file as input, to replace "load_word_topics" and "get_topics" in the future.

"load_word_topics" uses Mallet state file, which can be huge if Mallet is run with
a lot of regions and cells as input.

Mallet state files are quite big (even several hunderds of GB) and take quite a bit
of time to be written to disk when Mallet is run with a lot of regions and cells.
Getting count values for each region-topic pair is quite memory intensive in the
current code too.
Mallet region topics counts files are much smaller and don't require much post
processing and time to be written to disk.