Use --word-topic-counts-file instead of --output-state file when running Mallet.
Use --word-topic-counts-file instead of --output-state file when running Mallet.
The first generates directly the counts we need, with minimal post processing.
The state file also could take ages to write and read for datasets with a lot
of regions and could increase the runtime of topic modeling by 24 hours.

Changed `LDAMallet.train` to `LDMallet.run_mallet_topic_modeling` and made
it a static method.

Reading of the output produced by Mallet will be added in further commits.