Fix "create_fragment_matrix_from_fragments" in case last (few) region IDs do not overlap with any fragments.
Fix "create_fragment_matrix_from_fragments" in case last (few)
region IDs do not overlap with any fragments by setting sparse
matrix dimensions. Otherwise the sparse matrix could have to
small dimensions for the number of returned region IDs.
The same could happen for CBs if CBs are given that are not
in the fragments file.