Add `get_marker_regions_for_contrast()` as a much faster and memory efficient `markers()` replacement.
Add `get_marker_regions_for_contrast()` as a much faster and memory
efficient `markers()` replacement.

Calculations are done in chunks now, and does not require excessive
amounts of RAM anymore (TB) when the number of regions and/or cells
are high. For example 1 milion cells and 1 milion regions would only
require 75 GB of RAM compared with more than 3725 GB RAM with the old
code. As chunking is based on the value of `regions_chunk_size` (20000),
lowering it to e.g. 2000 would allow to process 100 milion cells with
only 745 GB of RAM.

`get_marker_regions_for_contrast()` is also much faster than `markers()`.

For example for a 1.73 milion cells dataset with 77k regions in foreground/background:
  - markers: 9320 seconds
  - get_marker_regions_for_contrast: 609 seconds